matching/glema/requirements.txt
matplotlib == 3.5.1
networkx == 3.1
pandas == 1.3.5
pillow == 8.4.0
scipy == 1.7.3
numpy == 1.21.2
tqdm == 4.62.3
pytorch == 1.8.2
scikit-learn == 0.24.2
ipykernel == 6.4.1